rename WordProcessor folder to WordProcessors because a class with the same name exists in it.
